Remove mostly everything not related to nodes
