--- Code/Include/BopAudio/Logging.h
@@ -1,5 +1,0 @@
-#pragma once
-
-namespace BopAudio
-{
-}
