--- Code/Include/BopAudio/ReflectionAttributes.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "AzCore/Math/Crc.h"
-
-namespace BopAudio::Attributes
-{
-    constexpr auto AudioTaskFactory{ AZ_CRC_CE("AudioTaskFactoryAttribute") };
-}
