+++ Code/Source/Clients/BopAudioSystemComponent.cpp
@@ -1,13 +1,9 @@
 #include "BopAudioSystemComponent.h"
 
-#include "AzCore/IO/FileIO.h"
 #include "AzCore/Serialization/SerializeContext.h"
 #include "AzCore/std/smart_ptr/unique_ptr.h"
-#include "Clients/AudioEventAsset.h"
 
 #include "BopAudio/BopAudioTypeIds.h"
-#include "Clients/AudioEventAssetHandler.h"
-#include "Engine/ConfigurationSettings.h"
 #include "ScriptCanvas/Nodes/AudioControlNode.h"
 
 namespace BopAudio
@@ -17,11 +13,6 @@ namespace BopAudio
 
     void BopAudioSystemComponent::Reflect(AZ::ReflectContext* context)
     {
-        StartEventData::Reflect(context);
-        StopEventData::Reflect(context);
-
-        AudioEventAsset::Reflect(context);
-
         if (auto* const serialize = azrtti_cast<AZ::SerializeContext*>(context))
         {
             serialize->Class<BopAudioSystemComponent, AZ::Component>()->Version(0);
@@ -32,14 +23,12 @@ namespace BopAudio
         AZ::ComponentDescriptor::DependencyArrayType& provided)
     {
         provided.push_back(AZ_CRC_CE("BopAudioService"));
-        provided.push_back(AZ_CRC_CE("AudioEngineService"));
     }
 
     void BopAudioSystemComponent::GetIncompatibleServices(
         AZ::ComponentDescriptor::DependencyArrayType& incompatible)
     {
         incompatible.push_back(AZ_CRC_CE("BopAudioService"));
-        incompatible.push_back(AZ_CRC_CE("AudioEngineService"));
     }
 
     void BopAudioSystemComponent::GetRequiredServices(
@@ -47,7 +36,6 @@ namespace BopAudio
     {
         required.push_back(AZ_CRC_CE("AssetDatabaseService"));
         required.push_back(AZ_CRC_CE("AssetCatalogService"));
-        required.push_back(AZ_CRC_CE("AudioSystemService"));
     }
 
     void BopAudioSystemComponent::GetDependentServices(
@@ -77,45 +65,14 @@ namespace BopAudio
         AZStd::make_unique<Nodes::AudioControlNode>();
     }
 
-    void BopAudioSystemComponent::RegisterFileAliases()
-    {
-        auto const banksPath = []()
-        {
-            static constexpr auto path{ "@products@/sounds/bopaudio/banks" };
-            return AZ::IO::FileIOBase::GetInstance()->ResolvePath(path).value_or("");
-        }();
-
-        auto const eventsPath = []()
-        {
-            static constexpr auto path{ "@products@/sounds/bopaudio/events" };
-            return AZ::IO::FileIOBase::GetInstance()->ResolvePath(path).value_or("");
-        }();
-
-        auto const projectPath = []()
-        {
-            static constexpr auto path{ "@products@/sounds/bopaudio" };
-            return AZ::IO::FileIOBase::GetInstance()->ResolvePath(path).value_or("");
-        }();
-
-        AZ::IO::FileIOBase::GetInstance()->SetAlias(BanksAlias, banksPath.c_str());
-        AZ::IO::FileIOBase::GetInstance()->SetAlias(EventsAlias, eventsPath.c_str());
-        AZ::IO::FileIOBase::GetInstance()->SetAlias(ProjectAlias, projectPath.c_str());
-    }
-
     void BopAudioSystemComponent::Activate()
     {
-        RegisterFileAliases();
-
-        m_audioEventAssetHandler.Register();
-
         BopAudioRequestBus::Handler::BusConnect();
     }
 
     void BopAudioSystemComponent::Deactivate()
     {
         BopAudioRequestBus::Handler::BusDisconnect();
-
-        m_audioEventAssetHandler.Unregister();
     }
 
 } // namespace BopAudio
