--- Code/Source/Engine/Rtpc.cpp
@@ -1,6 +1,0 @@
-#include "Engine/Rtpc.h"
-
-namespace BopAudio
-{
-
-}
