--- Code/Source/Engine/Rtpc.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace BopAudio
-{
-
-}
