--- Code/Source/Tools/BopAudio.qrc
@@ -1,6 +1,0 @@
-<!DOCTYPE RCC><RCC version="1.0">
-    <qresource prefix="/BopAudio">
-        <file alias="toolbar_icon.svg">toolbar_icon.svg</file>
-        <file alias="trigger_icon.svg">trigger_icon.svg</file>
-    </qresource>
-</RCC>
