+++ Code/Source/Tools/BopAudioEditorModule.cpp
@@ -1,15 +1,7 @@
 #include "BopAudioModuleInterface.h"
 
 #include "BopAudio/BopAudioTypeIds.h"
-#include "Builder/BopAudioAssetBuilderComponent.h"
 #include "Tools/BopAudioEditorSystemComponent.h"
-
-void InitBopAudioResources()
-{
-    // We must register our Qt resources (.qrc file) since this is being loaded
-    // from a separate module (gem)
-    Q_INIT_RESOURCE(BopAudio);
-}
 
 namespace BopAudio
 {
@@ -21,8 +13,6 @@ namespace BopAudio
 
         BopAudioEditorModule()
         {
-            InitBopAudioResources();
-
             // Push results of [MyComponent]::CreateDescriptor() into
             // m_descriptors here. Add ALL components descriptors associated
             // with this gem to m_descriptors. This will associate the
@@ -33,7 +23,6 @@ namespace BopAudio
                 m_descriptors.end(),
                 {
                     BopAudioEditorSystemComponent::CreateDescriptor(),
-                    BopAudioAssetBuilderComponent::CreateDescriptor(),
                 });
         }
