+++ Code/Source/Tools/BopAudioEditorSystemComponent.h
@@ -11,7 +11,6 @@ namespace BopAudio
     class BopAudioEditorSystemComponent
         : public BopAudioSystemComponent
         , protected AzToolsFramework::EditorEvents::Bus::Handler
-        , protected AudioControlsEditor::EditorImplPluginEventBus::Handler
     {
         using BaseSystemComponent = BopAudioSystemComponent;
 
@@ -34,18 +33,6 @@ namespace BopAudio
         void Init() override;
         void Activate() override;
         void Deactivate() override;
-
-    protected:
-        // AzToolsFramework::EditorEventsBus overrides ...
-        void NotifyRegisterViews() override;
-
-        ////////////////////////////////////////////////////////////////////////
-        // AudioControlsEditor::EditorImplPluginEventBus interface
-        // implementation
-        void InitializeEditorImplPlugin() override;
-        void ReleaseEditorImplPlugin() override;
-        auto GetEditorImplPlugin() -> AudioControls::IAudioSystemEditor* override;
-        ////////////////////////////////////////////////////////////////////////
 
     private:
         AZStd::unique_ptr<AudioControls::IAudioSystemEditor> m_editorImplPlugin;
