--- Code/Source/Tools/trigger_icon.source
@@ -1,1 +1,0 @@
-https://www.svgrepo.com/svg/449305/trigger
