+++ Code/Tests/Clients/BopAudioTest.cpp
@@ -1,7 +1,5 @@
-#include "AzCore/Asset/AssetTypeInfoBus.h"
 
-#include "Clients/AudioEventAsset.h"
-#include "Clients/BootstrapFixture.h"
+#include "Clients/BaseTestFixture.h"
 
 using ::testing::AtLeast;
 using ::testing::NiceMock;
@@ -11,13 +9,5 @@ namespace BopAudioTests
 {
     TEST_F(BaseAudioTestFixture, SANITY_CHECK)
     {
-    }
-
-    TEST_F(BootstrapFixture, AudioEventAssetAssetHandler_IsConnected)
-    {
-        EXPECT_EQ(
-            AZ::AssetTypeInfoBus::GetNumOfEventHandlers(
-                AZ::AzTypeInfo<BopAudio::AudioEventAsset>::Uuid()),
-            1);
     }
 } // namespace BopAudioTests
