+++ Code/bopaudio_api_files.cmake
@@ -3,7 +3,4 @@ set(FILES
     Include/BopAudio/BopAudioBus.h
     Include/BopAudio/BopAudioTypeIds.h
     Include/BopAudio/BopAudioComponentBus.h
-    Include/BopAudio/Util.h
-    Include/BopAudio/ReflectionAttributes.h
-    Include/BopAudio/Logging.h
 )
