+++ Code/bopaudio_editor_private_files.cmake
@@ -2,21 +2,4 @@
 set(FILES
     Source/Tools/BopAudioEditorSystemComponent.cpp
     Source/Tools/BopAudioEditorSystemComponent.h
-    Source/Tools/BopAudioWidget.cpp
-    Source/Tools/BopAudioWidget.h
-    Source/Tools/BopAudio.qrc
-
-    Source/Tools/AudioBopAudioLoader.cpp
-    Source/Tools/AudioBopAudioLoader.h
-    Source/Tools/AudioSystemControl_BopAudio.cpp
-    Source/Tools/AudioSystemControl_BopAudio.h
-    Source/Tools/AudioSystemEditor_BopAudio.cpp
-    Source/Tools/AudioSystemEditor_BopAudio.h
-
-    Source/Builder/AudioControlBuilderWorker.cpp
-    Source/Builder/AudioControlBuilderWorker.h
-    Source/Builder/AudioEventAssetBuilderWorker.cpp
-    Source/Builder/AudioEventAssetBuilderWorker.h
-    Source/Builder/BopAudioAssetBuilderComponent.cpp
-    Source/Builder/BopAudioAssetBuilderComponent.h
 )
