+++ Code/bopaudio_private_files.cmake
@@ -1,20 +1,11 @@
 
 set(FILES
-
     Source/BopAudioModuleInterface.cpp
     Source/BopAudioModuleInterface.h
     Source/Clients/BopAudioSystemComponent.cpp
     Source/Clients/BopAudioSystemComponent.h
+    
+    Source/Clients/ParameterTypes.cpp
+    Source/Clients/ParameterTypes.h
 
-    Source/Clients/AudioEventAsset.cpp
-    Source/Clients/AudioEventAsset.h
-    Source/Clients/AudioEventAssetHandler.cpp
-    Source/Clients/AudioEventAssetHandler.h
-
-    Source/Engine/Common_BopAudio.h
-    Source/Engine/ConfigurationSettings.cpp
-    Source/Engine/ConfigurationSettings.h
-    Source/Engine/DocumentReader.h
-    Source/Engine/Rtpc.cpp
-    Source/Engine/Rtpc.h
 )
